Remove obsolete SourceState class and name AbstractSourceState appropriately
