+++ src/main/java/org/janelia/saalfeldlab/paintera/control/Paint.java
@@ -19,7 +19,7 @@ import org.janelia.saalfeldlab.paintera.control.paint.SelectNextId;
 import org.janelia.saalfeldlab.paintera.control.selection.SelectedIds;
 import org.janelia.saalfeldlab.paintera.data.mask.CannotPersist;
 import org.janelia.saalfeldlab.paintera.data.mask.MaskedSource;
-import org.janelia.saalfeldlab.paintera.state.AbstractSourceState;
+import org.janelia.saalfeldlab.paintera.state.SourceState;
 import org.janelia.saalfeldlab.paintera.state.GlobalTransformManager;
 import org.janelia.saalfeldlab.paintera.state.LabelSourceState;
 import org.janelia.saalfeldlab.paintera.state.SourceInfo;
@@ -193,7 +193,7 @@ public class Paint implements ToOnEnterOnExit
 		};
 	}
 
-	public SelectedIds selectedIdsFromState( final AbstractSourceState< ?, ? > state )
+	public SelectedIds selectedIdsFromState( final SourceState< ?, ? > state )
 	{
 		return state instanceof LabelSourceState< ?, ? >
 				? ( ( LabelSourceState< ?, ? > ) state ).selectedIds()
