+++ src/main/java/org/janelia/saalfeldlab/paintera/control/paint/FloodFill.java
@@ -11,7 +11,7 @@ import java.util.function.Supplier;
 import org.janelia.saalfeldlab.paintera.data.mask.MaskInUse;
 import org.janelia.saalfeldlab.paintera.data.mask.MaskInfo;
 import org.janelia.saalfeldlab.paintera.data.mask.MaskedSource;
-import org.janelia.saalfeldlab.paintera.state.AbstractSourceState;
+import org.janelia.saalfeldlab.paintera.state.SourceState;
 import org.janelia.saalfeldlab.paintera.state.LabelSourceState;
 import org.janelia.saalfeldlab.paintera.state.SourceInfo;
 import org.slf4j.Logger;
@@ -93,7 +93,7 @@ public class FloodFill
 			return;
 		}
 
-		final AbstractSourceState< ?, ? > currentSourceState = sourceInfo.getState( currentSource );
+		final SourceState< ?, ? > currentSourceState = sourceInfo.getState( currentSource );
 
 		if ( !( currentSourceState instanceof LabelSourceState< ?, ? > ) )
 		{
