+++ src/main/java/org/janelia/saalfeldlab/paintera/control/paint/SelectNextId.java
@@ -5,7 +5,7 @@ import java.util.function.BiConsumer;
 
 import org.janelia.saalfeldlab.paintera.control.selection.SelectedIds;
 import org.janelia.saalfeldlab.paintera.id.IdService;
-import org.janelia.saalfeldlab.paintera.state.AbstractSourceState;
+import org.janelia.saalfeldlab.paintera.state.SourceState;
 import org.janelia.saalfeldlab.paintera.state.LabelSourceState;
 import org.janelia.saalfeldlab.paintera.state.SourceInfo;
 import org.slf4j.Logger;
@@ -47,7 +47,7 @@ public class SelectNextId
 			return;
 		}
 
-		final AbstractSourceState< ?, ? > currentSourceState = sourceInfo.getState( currentSource );
+		final SourceState< ?, ? > currentSourceState = sourceInfo.getState( currentSource );
 
 		if ( !( currentSourceState instanceof LabelSourceState< ?, ? > ) )
 		{
